Setup standard Devise user
The previous approach of creating an Achiever as a Devise user was a bit
misguided - the achiever model should only hold logic related to it.

Let's leave the User class to handle the cookie cutter authentication
data and logic. The achiever will be tied to the user in some way.

Also this has the benefit of being able to simply have users that have
more than 1 profile (achiever).